Minor test fix
