+++ aptos-move/aptos-vm-types/src/tests/utils.rs
@@ -52,23 +52,6 @@ pub(crate) fn raw_metadata(v: u64) -> StateValueMetadata {
     StateValueMetadata::new(v, &CurrentTimeMicroseconds { microseconds: v })
 }
 
-pub(crate) fn write_op_with_metadata(type_idx: u8, v: u128) -> WriteOp {
-    match type_idx {
-        0 => WriteOp::CreationWithMetadata {
-            data: as_bytes!(v).into(),
-            metadata: raw_metadata(v as u64),
-        },
-        1 => WriteOp::ModificationWithMetadata {
-            data: as_bytes!(v).into(),
-            metadata: raw_metadata(v as u64),
-        },
-        2 => WriteOp::DeletionWithMetadata {
-            metadata: raw_metadata(v as u64),
-        },
-        _ => unreachable!("Wrong type index for test"),
-    }
-}
-
 pub(crate) fn mock_create(k: impl ToString, v: u128) -> (StateKey, WriteOp) {
     (as_state_key!(k), WriteOp::Creation(as_bytes!(v).into()))
 }
